+++ src/streamwriter.h
@@ -50,7 +50,7 @@ namespace rcp {
         virtual void write(const rcp::Color& s);
         virtual void write(const rcp::IPv4& s);
         virtual void write(const rcp::IPv6& s);
-        virtual void write(const char* data, uint32_t length);
+        virtual void write(const char* data, size_t length);
 
     private:
         std::ostream& os;
